LayoutTests/imported/w3c/web-platform-tests/css/css-variables/variable-css-wide-keywords-expected.txt

`initial` as a value for an unregistered custom property
`inherit` as a value for an unregistered custom property
`unset` as a value for an unregistered custom property
`revert` as a value for an unregistered custom property
`revert-layer` as a value for an unregistered custom property
`initial` as a value for a non-inheriting registered custom property
`initial` as a value for an inheriting registered custom property
`inherit` as a value for a non-inheriting registered custom property
`inherit` as a value for an inheriting registered custom property
`unset` as a value for a non-inheriting registered custom property
`unset` as a value for an inheriting registered custom property
`revert` as a value for a non-inheriting registered custom property
`revert` as a value for an inheriting registered custom property
`revert-layer` as a value for a non-inheriting registered custom property
`revert-layer` as a value for an inheriting registered custom property
`initial` as a `var()` fallback for an unregistered custom property
`inherit` as a `var()` fallback for an unregistered custom property
`unset` as a `var()` fallback for an unregistered custom property
`revert` as a `var()` fallback for an unregistered custom property
`revert-layer` as a `var()` fallback for an unregistered custom property
`initial` as a `var()` fallback for a non-inheriting registered custom property
`initial` as a `var()` fallback for an inheriting registered custom property
`inherit` as a `var()` fallback for a non-inheriting registered custom property
`inherit` as a `var()` fallback for an inheriting registered custom property
`unset` as a `var()` fallback for a non-inheriting registered custom property
`unset` as a `var()` fallback for an inheriting registered custom property
`revert` as a `var()` fallback for a non-inheriting registered custom property
`revert` as a `var()` fallback for an inheriting registered custom property
`revert-layer` as a `var()` fallback for a non-inheriting registered custom property
`revert-layer` as a `var()` fallback for an inheriting registered custom property

PASS `initial` as a value for an unregistered custom property
PASS `inherit` as a value for an unregistered custom property
PASS `unset` as a value for an unregistered custom property
PASS `revert` as a value for an unregistered custom property
PASS `revert-layer` as a value for an unregistered custom property
PASS `initial` as a value for a non-inheriting registered custom property
PASS `initial` as a value for an inheriting registered custom property
PASS `inherit` as a value for a non-inheriting registered custom property
PASS `inherit` as a value for an inheriting registered custom property
PASS `unset` as a value for a non-inheriting registered custom property
PASS `unset` as a value for an inheriting registered custom property
PASS `revert` as a value for a non-inheriting registered custom property
PASS `revert` as a value for an inheriting registered custom property
PASS `revert-layer` as a value for a non-inheriting registered custom property
PASS `revert-layer` as a value for an inheriting registered custom property
FAIL `initial` as a `var()` fallback for an unregistered custom property assert_equals: expected "rgb(144, 238, 144)" but got "rgba(0, 0, 0, 0)"
FAIL `inherit` as a `var()` fallback for an unregistered custom property assert_equals: expected "rgb(144, 238, 144)" but got "rgba(0, 0, 0, 0)"
FAIL `unset` as a `var()` fallback for an unregistered custom property assert_equals: expected "rgb(144, 238, 144)" but got "rgba(0, 0, 0, 0)"
FAIL `revert` as a `var()` fallback for an unregistered custom property assert_equals: expected "rgb(144, 238, 144)" but got "rgba(0, 0, 0, 0)"
FAIL `revert-layer` as a `var()` fallback for an unregistered custom property assert_equals: expected "rgb(144, 238, 144)" but got "rgba(0, 0, 0, 0)"
PASS `initial` as a `var()` fallback for a non-inheriting registered custom property
PASS `initial` as a `var()` fallback for an inheriting registered custom property
PASS `inherit` as a `var()` fallback for a non-inheriting registered custom property
PASS `inherit` as a `var()` fallback for an inheriting registered custom property
PASS `unset` as a `var()` fallback for a non-inheriting registered custom property
PASS `unset` as a `var()` fallback for an inheriting registered custom property
PASS `revert` as a `var()` fallback for a non-inheriting registered custom property
PASS `revert` as a `var()` fallback for an inheriting registered custom property
PASS `revert-layer` as a `var()` fallback for a non-inheriting registered custom property
PASS `revert-layer` as a `var()` fallback for an inheriting registered custom property
